feat(playground): refactor sidebar configuration and improve playground page structure
