+++ docs/.vitepress/config.mts
@@ -2,14 +2,8 @@ import { fileURLToPath } from 'node:url'
 import { resolve, dirname } from 'node:path'
 import tailwindcss from '@tailwindcss/vite'
 import { defineConfigWithTheme } from 'vitepress'
-import { useSidebar } from 'vitepress-openapi'
-import { examplesPages, testsPages } from '../pages'
-import spec from '../public/openapi.json' with {type: 'json'}
+import { createDocsSidebar } from '../docs-sidebar-config'
 import { sidebarExamplesConfigs } from '../sidebar-examples-configs'
-
-const sidebar = useSidebar({
-  spec,
-})
 
 const gaId = process.env.GA_ID || 'G-TEST'
 
@@ -37,240 +31,7 @@ export default defineConfigWithTheme({
 
     sidebar: {
       ...sidebarExamplesConfigs,
-      '/': [
-        {
-          text: 'Getting Started',
-          link: '/guide/getting-started',
-        },
-        {
-          text: 'Pages',
-          items: [
-            {
-              items: [
-                {
-                  text: 'By Operation',
-                  link: '/pages/by-operation',
-                },
-                {
-                  text: 'By Spec',
-                  link: '/pages/by-spec',
-                },
-                {
-                  text: 'By Tag',
-                  link: '/pages/by-tag',
-                },
-                {
-                  text: 'Introduction',
-                  link: '/pages/introduction',
-                },
-              ],
-            },
-          ],
-        },
-        {
-          text: 'Sidebar',
-          items: [
-            {
-              items: [
-                {
-                  text: 'Sidebar Items',
-                  link: '/sidebar/sidebar-items',
-                },
-                {
-                  text: 'Sidebar Examples',
-                  link: '/sidebar-examples/',
-                  target: '_blank',
-                },
-              ],
-            },
-          ],
-        },
-        {
-          text: 'Components',
-          items: [
-            {
-              items: [
-                {
-                  text: 'Overview',
-                  link: '/components/',
-                },
-                {
-                  text: 'OASpec',
-                  link: '/components/oa-spec',
-                },
-                {
-                  text: 'OASpecPlayground',
-                  link: '/components/oa-spec-playground',
-                },
-                {
-                  text: 'OAOperation',
-                  link: '/components/oa-operation',
-                },
-              ],
-            },
-          ],
-        },
-        {
-          text: 'Composables',
-          items: [
-            {
-              items: [
-                {
-                  text: 'useTheme',
-                  link: '/composables/useTheme',
-                },
-                {
-                  text: 'usePlayground',
-                  link: '/composables/usePlayground',
-                },
-              ],
-            },
-          ],
-        },
-        {
-          text: 'Utilities',
-          items: [
-            {
-              items: [
-                {
-                  text: 'minifyHtml',
-                  link: '/utils/minifyHtml',
-                },
-              ],
-            },
-          ],
-        },
-        {
-          text: 'Customizations',
-          items: [
-            {
-              items: [
-                {
-                  text: 'General',
-                  items: [
-                    {
-                      items: [
-                        {
-                          text: 'i18n',
-                          link: '/customizations/i18n',
-                        },
-                        {
-                          text: 'Branding',
-                          link: '/customizations/branding',
-                        },
-                      ],
-                    },
-                  ],
-                },
-                {
-                  text: 'Operation',
-                  items: [
-                    {
-                      items: [
-                        {
-                          text: 'Custom Slots',
-                          link: '/customizations/custom-slots',
-                        },
-                        {
-                          text: 'Operation Badges',
-                          link: '/customizations/operation-badges',
-                        },
-                        {
-                          text: 'Code Samples',
-                          link: '/customizations/code-samples',
-                        },
-                        {
-                          text: 'Operation tags slot',
-                          link: '/customizations/operation-tags-slot',
-                        },
-                        {
-                          text: 'Custom Server',
-                          link: '/customizations/custom-server',
-                        },
-                        {
-                          text: 'Operation Links',
-                          link: '/customizations/operation-links',
-                        },
-                      ],
-                    },
-                  ],
-                },
-                {
-                  text: 'Spec',
-                  items: [
-                    {
-                      items: [
-                        {
-                          text: 'Multiple Specs',
-                          link: '/customizations/multiple-specs',
-                        },
-                      ],
-                    },
-                  ],
-                },
-                {
-                  text: 'Playground',
-                  items: [
-                    {
-                      items: [
-                        {
-                          text: 'Examples',
-                          link: '/customizations/playground-examples',
-                        },
-                      ],
-                    },
-                  ],
-                },
-              ],
-            },
-          ],
-        },
-        {
-          text: 'Example',
-          collapsed: true,
-          items: [
-            {
-              text: 'Introduction',
-              link: '/example/introduction',
-            },
-            ...sidebar.generateSidebarGroups({
-              linkPrefix: '/example/operations/',
-            }),
-            {
-              text: 'One Page',
-              link: '/example/one-page',
-            },
-            {
-              text: 'Playground',
-              link: '/example/playground',
-            },
-            {
-              text: 'Playground Custom Sidebar',
-              link: '/example/playground-custom-sidebar',
-            },
-          ],
-        },
-        {
-          text: 'Tests',
-          collapsed: true,
-          items: [
-            ...testsPages.map(page => ({
-              text: page.label,
-              link: `/tests/${page.slug}`,
-            })),
-          ],
-        },
-        {
-          text: 'Remote Examples',
-          collapsed: true,
-          items: [
-            ...examplesPages.map(page => ({
-              text: page.label,
-              link: `/examples/${page.slug}`,
-            })),
-          ],
-        },
-      ],
+      '/': createDocsSidebar(),
     },
 
     socialLinks: [
