+++ docs/.vitepress/theme/style.css
@@ -8,3 +8,15 @@
 .VPDoc .content .content-container {
   max-width: none !important;
 }
+
+.VPDoc:has(.OASpecPlayground) .OASpecPlayground .OASidebar,
+.VPDoc:has(.OASpecPlayground) .OASpecPlayground .OAPlaygroundLocalNav,
+.VPDoc:has(.OASpecPlayground) .OASpecPlayground .VPBackdrop {
+  display: none;
+}
+
+.VPDoc:has(.OASpecPlayground) .OASpecPlayground .OAContent.has-sidebar {
+  margin-top: 0;
+  padding-left: 0;
+  padding-right: 0;
+}
